+++ ProyectoParkingServices/RegistroPlaza.cs
@@ -8,7 +8,7 @@ namespace ProyectoParkingServices
         public int Id { get; set; }
         public int Id_Plaza { get; set; }
         public int Id_Coche { get; set; }
-        public int Id_Parking {  get; set; }
+        public Int16 Id_Parking {  get; set; }
         public Boolean tipoEvento { get; set; }
         public DateTime fechaEvento {get; set;}
